refact: evm test page
