+++ packages/example/components/ethereum/styles.module.scss
@@ -1,0 +1,77 @@
+.main {
+  section {
+    margin: 20px 0 20px 0;
+  }
+  
+  :global(.info-text) {
+    font-size: 1.1em;
+    word-break: break-word;
+  }
+  
+  :global(.full-width) {
+    width: 100%;
+  }
+  
+  :global(.card) {
+    margin-bottom: 20px;
+  }
+  
+  /* Logo & Header */
+  
+  header {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+  
+  :global(#logo-container) {
+    padding-top: 2rem;
+    display: inline-flex;
+    flex-direction: column;
+    align-items: center;
+  }
+  
+  :global(#logo-text) {
+    padding-top: 15px;
+    padding-bottom: 5px;
+  }
+  
+  :global(#logo) {
+    width: 65px;
+    height: 66px;
+  }
+  
+  /* Encrypt/Decrypt */
+  
+  :global(#encryptMessageInput) {
+    margin-bottom: 1rem;
+  }
+  
+  :global(.error-div) {
+    margin-bottom: 12px;
+  }
+  
+  :global(.error-message) {
+    min-height: 32px;
+    border: 1px solid #e88f97;
+    color: #24292e;
+    background: #fcf2f3;
+    border-radius: 8px;
+    display: flex;
+    justify-content: flex-start;
+    align-items: center;
+    padding: 8px 10px;
+  }
+  :global(.error-message-icon) {
+    margin-right: 8px;
+    flex: 0 0 auto;
+  }
+  :global(.error-message-text) {
+    overflow: auto;
+  }
+  
+  :global(.warning-invisible) {
+    display: none;
+  }
+  
+}
