+++ packages/example/package.json
@@ -10,8 +10,12 @@
     "clean": "rm -rf ./node_modules && rm -rf ./.next && rm -rf ./out"
   },
   "dependencies": {
+    "@metamask/onboarding": "^1.0.1",
     "@onekeyfe/onekey-near-provider": "^0.0.9",
     "console-feed": "^3.3.0",
+    "eth-sig-util": "^3.0.1",
+    "ethereumjs-util": "^7.1.4",
+    "ethers": "^5.6.5",
     "near-api-js": "^0.44.2",
     "next": "^12.1.0",
     "react": "17.0.2",
@@ -23,6 +27,7 @@
     "@types/react": "17.0.39",
     "eslint": "8.9.0",
     "eslint-config-next": "12.0.10",
+    "sass": "^1.51.0",
     "typescript": "4.5.5"
   }
 }
