--- packages/example/public/evm/index.css
@@ -1,74 +1,0 @@
-section {
-  margin: 20px 0 20px 0;
-}
-
-.info-text {
-  font-size: 1.1em;
-  word-break: break-word;
-}
-
-.full-width {
-  width: 100%;
-}
-
-.card {
-  margin-bottom: 20px;
-}
-
-/* Logo & Header */
-
-header {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-#logo-container {
-  padding-top: 2rem;
-  display: inline-flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-#logo-text {
-  padding-top: 15px;
-  padding-bottom: 5px;
-}
-
-#logo {
-  width: 65px;
-  height: 66px;
-}
-
-/* Encrypt/Decrypt */
-
-#encryptMessageInput {
-  margin-bottom: 1rem;
-}
-
-.error-div {
-  margin-bottom: 12px;
-}
-
-.error-message {
-  min-height: 32px;
-  border: 1px solid #e88f97;
-  color: #24292e;
-  background: #fcf2f3;
-  border-radius: 8px;
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  padding: 8px 10px;
-}
-.error-message-icon {
-  margin-right: 8px;
-  flex: 0 0 auto;
-}
-.error-message-text {
-  overflow: auto;
-}
-
-.warning-invisible {
-  display: none;
-}
