+++ packages/example/public/logo.png
